Move validation code to DeFi
